remove `var_int` and `var_long` and merge into the root of the library
